+++ src/libs/Detach.ImGuiUtilities/ImPlotFormatterDelegates.cs
@@ -1,4 +1,4 @@
-﻿namespace Detach.ImGuiUtilities.Utils;
+﻿namespace Detach.ImGuiUtilities;
 
 public static unsafe class ImPlotFormatterDelegates
 {